+++ packages/apm/.npmignore
@@ -1,0 +1,4 @@
+*
+!/dist/**/*
+!/esm/**/*
+*.tsbuildinfo
