+++ packages/apm/package.json
@@ -1,49 +1,45 @@
 {
-  "name": "@sentry/integrations",
-  "version": "5.8.0",
-  "description": "Pluggable integrations that can be used to enchance JS SDKs",
+  "name": "@sentry/apm",
+  "version": "5.10.0-beta.3",
+  "description": "Extensions for APM",
   "repository": "git://github.com/getsentry/sentry-javascript.git",
-  "homepage": "https://github.com/getsentry/sentry-javascript/tree/master/packages/integrations",
+  "homepage": "https://github.com/getsentry/sentry-javascript/tree/master/packages/core",
   "author": "Sentry",
   "license": "BSD-3-Clause",
   "engines": {
     "node": ">=6"
   },
-  "publishConfig": {
-    "access": "public"
-  },
   "main": "dist/index.js",
   "module": "esm/index.js",
   "types": "dist/index.d.ts",
+  "publishConfig": {
+    "access": "public"
+  },
   "dependencies": {
-    "@sentry/types": "5.7.1",
-    "@sentry/utils": "5.8.0",
+    "@sentry/hub": "5.10.0-beta.3",
+    "@sentry/minimal": "5.10.0-beta.3",
+    "@sentry/types": "5.10.0-beta.3",
+    "@sentry/utils": "5.10.0-beta.3",
     "tslib": "^1.9.3"
   },
   "devDependencies": {
-    "chai": "^4.1.2",
+    "@types/express": "^4.17.1",
     "jest": "^24.7.1",
     "npm-run-all": "^4.1.2",
     "prettier": "^1.17.0",
     "prettier-check": "^2.0.0",
     "rimraf": "^2.6.3",
-    "rollup": "^1.10.1",
-    "rollup-plugin-commonjs": "^9.3.4",
-    "rollup-plugin-node-resolve": "^4.2.3",
-    "rollup-plugin-terser": "^4.0.4",
-    "rollup-plugin-typescript2": "^0.21.0",
     "tslint": "^5.16.0",
     "typescript": "^3.4.5"
   },
   "scripts": {
-    "build": "run-p build:es5 build:esm build:bundle",
+    "build": "run-p build:es5 build:esm",
     "build:es5": "tsc -p tsconfig.build.json",
     "build:esm": "tsc -p tsconfig.esm.json",
     "build:watch": "run-p build:watch:es5 build:watch:esm",
     "build:watch:es5": "tsc -p tsconfig.build.json -w --preserveWatchOutput",
     "build:watch:esm": "tsc -p tsconfig.esm.json -w --preserveWatchOutput",
-    "build:bundle": "rollup --config",
-    "clean": "rimraf dist coverage esm build .rpt2_cache",
+    "clean": "rimraf dist coverage",
     "link:yarn": "yarn link",
     "lint": "run-s lint:prettier lint:tslint",
     "lint:prettier": "prettier-check \"{src,test}/**/*.ts\"",
