+++ packages/apm/src/index.ts
@@ -1,0 +1,5 @@
+import * as ApmIntegrations from './integrations';
+
+export { Hub } from './hub';
+export { ApmIntegrations as Integrations };
+export { Span, TRACEPARENT_REGEXP } from './span';
