+++ packages/apm/src/integrations/index.ts
@@ -1,0 +1,3 @@
+export { Express } from './express';
+export { Tracing } from './tracing';
+export { TransactionActivity, TransactionActivityHandlers } from './transactionactivity';
