+++ packages/apm/src/integrations/tracing.ts
@@ -1,5 +1,8 @@
-import { EventProcessor, Hub, Integration } from '@sentry/types';
+import { makeMain } from '@sentry/hub';
+import { EventProcessor, Integration } from '@sentry/types';
 import { fill, getGlobalObject, isMatchingPattern, logger, supportsNativeFetch } from '@sentry/utils';
+
+import { Hub } from '../hub';
 
 /** JSDoc */
 interface TracingOptions {
@@ -35,6 +38,7 @@ export class Tracing implements Integration {
    * @param _options TracingOptions
    */
   public constructor(private readonly _options: TracingOptions = {}) {
+    makeMain(new Hub());
     if (!Array.isArray(_options.tracingOrigins) || _options.tracingOrigins.length === 0) {
       const defaultTracingOrigins = ['localhost', /^\//];
       logger.warn(
