+++ packages/apm/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@sentry/typescript/tslint"
+}
