+++ packages/hub/src/index.ts
@@ -1,4 +1,3 @@
 export { Carrier, Layer } from './interfaces';
 export { addGlobalEventProcessor, Scope } from './scope';
 export { getCurrentHub, getHubFromCarrier, getMainCarrier, Hub, makeMain, setHubOnCarrier } from './hub';
-export { Span, TRACEPARENT_REGEXP } from './span';
