+++ packages/integrations/src/index.ts
@@ -3,12 +3,9 @@ export { CaptureConsole } from './captureconsole';
 export { Debug } from './debug';
 export { Dedupe } from './dedupe';
 export { Ember } from './ember';
-export { Express } from './express';
 export { ExtraErrorData } from './extraerrordata';
 export { ReportingObserver } from './reportingobserver';
 export { RewriteFrames } from './rewriteframes';
 export { SessionTiming } from './sessiontiming';
-export { Tracing } from './tracing';
 export { Transaction } from './transaction';
-export { TransactionActivity, TransactionActivityHandlers } from './transactionactivity';
 export { Vue } from './vue';
