+++ packages/types/src/hub.ts
@@ -4,7 +4,6 @@ import { Event, EventHint } from './event';
 import { Integration, IntegrationClass } from './integration';
 import { Scope } from './scope';
 import { Severity } from './severity';
-import { Span, SpanContext } from './span';
 import { User } from './user';
 
 /**
@@ -168,16 +167,4 @@ export interface Hub {
 
   /** Returns the integration if installed on the current client. */
   getIntegration<T extends Integration>(integration: IntegrationClass<T>): T | null;
-
-  /** Returns all trace headers that are currently on the top scope. */
-  traceHeaders(): { [key: string]: string };
-
-  /**
-   * This functions starts a span. If argument passed is of type `Span`, it'll run sampling on it if configured
-   * and attach a `SpanRecorder`. If it's of type `SpanContext` and there is already a `Span` on the Scope,
-   * the created Span will have a reference to it and become it's child. Otherwise it'll crete a new `Span`.
-   *
-   * @param span Already constructed span which should be started or properties with which the span should be created
-   */
-  startSpan(span?: Span | SpanContext, forceNoChild?: boolean): Span;
 }
